Fix Makefiles for wasm of router and reverse-echo and mandlebrot
