+++ samples/reverse-echo/Makefile
@@ -6,4 +6,4 @@ src/$(NAME).wasm: target/wasm32-unknown-unknown/debug/$(NAME).wasm
 	@cp $< $@
 
 target/wasm32-unknown-unknown/debug/$(NAME).wasm: src/$(NAME).rs Cargo.toml
-	cargo build --lib --target=wasm32-unknown-unknown
+	@cargo build --quiet --lib --target=wasm32-unknown-unknown
