+++ samples/router/context.toml
@@ -10,7 +10,7 @@ to = "process/read_section"
 
 [[process]]
 alias = "read_section"
-source = "read_section.toml"
+source = "src/read_section.toml"
 
 [[connection]]
 from = "process/read_section"
